Move `IndexAnnotationBuilder` => `IndexAnnotation::AnnotationBuilder`
